backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pillow==10.0.0
pytesseract==0.3.10
aiofiles==23.1.0

passlib[bcrypt]==1.7.4
pymongo[srv]==4.7.3
python-dotenv==1.0.0
